setup elastic search

+++ Domain/Model/CollectionMatch.cs
@@ -7,7 +7,7 @@ namespace Domain.Model
         public string ExportDate { get; set; }
 
         [ForeignKey("ArtistCollection")]
-		public int CollectionId { get; set; }
+		public long CollectionId { get; set; }
 		public string Upc { get; set; }
 		public string Grid { get; set; }
 		public string AmgAlbumId { get; set; }
